+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -45,6 +45,8 @@ import mozilla.components.concept.engine.mediasession.MediaSession
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.search.SearchRequest
+import mozilla.components.concept.engine.translate.TranslationOperation
+import mozilla.components.concept.engine.translate.TranslationOptions
 import mozilla.components.concept.engine.webextension.WebExtensionBrowserAction
 import mozilla.components.concept.engine.webextension.WebExtensionPageAction
 import mozilla.components.concept.engine.window.WindowRequest
@@ -835,6 +837,59 @@ sealed class ContentAction : BrowserAction() {
         val tabId: String,
         val isProductUrl: Boolean,
     ) : ContentAction()
+}
+
+/**
+ * [BrowserAction] implementations related to translating a web content page.
+ */
+sealed class TranslationsAction : BrowserAction() {
+    /**
+     * Used to translate the page for a given [tabId].
+     *
+     * @property tabId The ID of the tab the [EngineSession] should be linked to.
+     * @property fromLanguage The BCP 47 language tag that the page should be translated from.
+     * @property toLanguage The BCP 47 language tag that the page should be translated to.
+     * @property options Options for how the translation should be processed.
+     */
+    data class TranslateAction(
+        override val tabId: String,
+        val fromLanguage: String,
+        val toLanguage: String,
+        val options: TranslationOptions?,
+    ) : TranslationsAction(), ActionWithTab
+
+    /**
+     * Indicates the given [tabId] should restore the original pre-translated content.
+     *
+     * @property tabId The ID of the tab the [EngineSession] should be linked to.
+     */
+    data class TranslateRestoreAction(
+        override val tabId: String,
+    ) : TranslationsAction(), ActionWithTab
+
+    /**
+     * Indicates the given [tabId] was successful in translating or restoring the page.
+     *
+     * @property tabId The ID of the tab the [EngineSession] should be linked to.
+     * @property operation The translation operation that was successful.
+     */
+    data class TranslateSuccessAction(
+        override val tabId: String,
+        val operation: TranslationOperation,
+    ) : TranslationsAction(), ActionWithTab
+
+    /**
+     * Indicates the given [tabId] was unable to translate or restore the page.
+     *
+     * @property tabId The ID of the tab the [EngineSession] should be linked to.
+     * @property operation The translation operation that failed.
+     * @property throwable The throwable for error handling.
+     */
+    data class TranslateExceptionAction(
+        override val tabId: String,
+        val operation: TranslationOperation,
+        val throwable: Throwable,
+    ) : TranslationsAction(), ActionWithTab
 }
 
 /**