+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -14,6 +14,7 @@ import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.action.MediaSessionAction
 import mozilla.components.browser.state.action.ReaderAction
 import mozilla.components.browser.state.action.TrackingProtectionAction
+import mozilla.components.browser.state.action.TranslationsAction
 import mozilla.components.browser.state.selector.findTabOrCustomTab
 import mozilla.components.browser.state.state.AppIntentState
 import mozilla.components.browser.state.state.BrowserState
@@ -32,6 +33,7 @@ import mozilla.components.concept.engine.media.RecordingDevice
 import mozilla.components.concept.engine.mediasession.MediaSession
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
+import mozilla.components.concept.engine.translate.TranslationOperation
 import mozilla.components.concept.engine.window.WindowRequest
 import mozilla.components.concept.fetch.Response
 import mozilla.components.lib.state.Store
@@ -463,5 +465,13 @@ internal class EngineObserver(
 
     override fun onCheckForFormDataException(throwable: Throwable) {
         store.dispatch(ContentAction.CheckForFormDataExceptionAction(tabId, throwable))
+    }
+
+    override fun onTranslateComplete(operation: TranslationOperation) {
+        store.dispatch(TranslationsAction.TranslateSuccessAction(tabId, operation))
+    }
+
+    override fun onTranslateException(operation: TranslationOperation, throwable: Throwable) {
+        store.dispatch(TranslationsAction.TranslateExceptionAction(tabId, operation, throwable))
     }
 }