+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/SessionState.kt
@@ -15,6 +15,7 @@ import mozilla.components.support.utils.SafeIntent
  * @property id the unique id of the session.
  * @property content the [ContentState] of this session.
  * @property trackingProtection the [TrackingProtectionState] of this session.
+ * @property translationsState the [TranslationsState] of this session.
  * @property cookieBanner Indicates the state of cookie banner for this session.
  * @property engineState the [EngineState] of this session.
  * @property extensionState a map of extension id and web extension states
@@ -29,6 +30,7 @@ interface SessionState {
     val id: String
     val content: ContentState
     val trackingProtection: TrackingProtectionState
+    val translationsState: TranslationsState
     val cookieBanner: CookieBannerHandlingStatus
     val engineState: EngineState
     val extensionState: Map<String, WebExtensionState>
@@ -45,6 +47,7 @@ interface SessionState {
         id: String = this.id,
         content: ContentState = this.content,
         trackingProtection: TrackingProtectionState = this.trackingProtection,
+        translationsState: TranslationsState = this.translationsState,
         engineState: EngineState = this.engineState,
         extensionState: Map<String, WebExtensionState> = this.extensionState,
         mediaSessionState: MediaSessionState? = this.mediaSessionState,